download files functionality

+++ src/main/java/com/sashkomusic/downloadagent/domain/model/DownloadOption.java
@@ -4,20 +4,20 @@ import java.util.List;
 import java.util.Map;
 
 public record DownloadOption(
-        String id,                 // Унікальний ID (для UI/ідентифікації)
-        String sourceName,         // Назва джерела (напр. "Soulseek")
-        String distributorName,    // Ім'я користувача (напр. "waxwax")
-        int totalSize,            // Загальний розмір релізу
-        List<FileItem> files,      // Список файлів всередині
-        Map<String, String> technicalMetadata // Технічні дані для скачування (token, username)
+        String id,
+        String sourceName,
+        String distributorName,
+        int totalSize,
+        List<FileItem> files,
+        Map<String, String> technicalMetadata
 ) {
     public record FileItem(
             String filename,
-            int size,
-            Integer bitRate,    // Може бути null (для FLAC)
-            Integer bitDepth,   // Може бути null (для MP3)
-            Integer sampleRate, // Може бути null
-            int lengthSeconds   // Тривалість
+            long size,
+            Integer bitRate,
+            Integer bitDepth,
+            Integer sampleRate,
+            int lengthSeconds
     ) {
     }
 }
